requirements.txt
supabase==2.7.4
requests
python-dotenv
schedule
pandas
